Organize folders and files
R material related to the html rendering put inside "resources" folder
