--- Samples/Sample_metadata.txt
@@ -1,16 +1,0 @@
-BMP_ID	target	Origin	Sample_name	Biome	latitude	longitude	Brazilian State	R1	R2	Alpha_div	Org.Matter	Average precip.		
-bmp1	16S	soil	J1	Cerrado	-21.226336	-44.981787	tomato	J01_S45_L001_R1_001.fastq.gz	J01_S45_L001_R2_001.fastq.gz	3.08	10	70		
-bmp2	16S	soil	J2	Cerrado	-21.226336	-44.981787	tomato	J02_S46_L001_R1_001.fastq.gz	J02_S46_L001_R2_001.fastq.gz	2.94	8.6	75		
-bmp3	16S	soil	J3	Cerrado	-21.226336	-44.981787	tomato	J03_S47_L001_R1_001.fastq.gz	J03_S47_L001_R2_001.fastq.gz	2.88	12	60		
-bmp4	16S	soil	J4	Cerrado	-15.608155	-56.062962	tomato	J04_S48_L001_R1_001.fastq.gz	J04_S48_L001_R2_001.fastq.gz	1.55	5	88		
-bmp5	16S	soil	J5	Cerrado	-15.608155	-56.062962	tomato	J05_S49_L001_R1_001.fastq.gz	J05_S49_L001_R2_001.fastq.gz	1.73	5.9	55		
-bmp6	16S	soil	J6	Cerrado	-15.608155	-56.062962	tomato	J06_S50_L001_R1_001.fastq.gz	J06_S50_L001_R2_001.fastq.gz	2	6.1	79		
-bmp7	16S	soil	J7	Cerrado	5.116982	-60.356531	pepper	J07_S51_L001_R1_001.fastq.gz	J07_S51_L001_R2_001.fastq.gz	0.5	1.5	80		
-bmp8	16S	soil	J8	Cerrado	5.116982	-60.356531	pepper	J08_S52_L001_R1_001.fastq.gz	J08_S52_L001_R2_001.fastq.gz	0.75	2	59		
-bmp9	16S	soil	J9	Cerrado	5.116982	-60.356531	pepper	J09_S53_L001_R1_001.fastq.gz	J09_S53_L001_R2_001.fastq.gz	0.6	1.8	62		
-bmp10	16S	soil	J10	Atlanctic Forest	-30.332433	-54.362428	pepper	J10_S54_L001_R1_001.fastq.gz	J10_S54_L001_R2_001.fastq.gz	3.2	5.5	120		
-bmp11	16S	soil	J11	Atlanctic Forest	-30.332433	-54.362428	pepper	J11_S55_L001_R1_001.fastq.gz	J11_S55_L001_R2_001.fastq.gz	3.3	6.1	128		
-bmp12	16S	soil	J12	Atlanctic Forest	-30.332433	-54.362428	pepper	J12_S56_L001_R1_001.fastq.gz	J12_S56_L001_R2_001.fastq.gz	3	5.8	150		
-bmp13	16S	soil	J13	Caatinga	-10.314699	-46.044695	tomato	J13_S57_L001_R1_001.fastq.gz	J13_S57_L001_R2_001.fastq.gz	2.5	11	35		
-bmp14	16S	soil	J14	Caatinga	-10.314699	-46.044695	tomato	J14_S58_L001_R1_001.fastq.gz	J14_S58_L001_R2_001.fastq.gz	1.8	10.9	30		
-bmp15	16S	soil	J15	Caatinga	-10.314699	-46.044695	tomato	J15_S59_L001_R1_001.fastq.gz	J15_S59_L001_R2_001.fastq.gz	2.3	9.7	22
