--- Samples/new_samps_metadata.txt
@@ -1,93 +1,0 @@
-Samples	Latitude	Longitude	Biome	Average.precip	Alpha_div	Org.Matter
-Sample_1	-19.56	-43.75	BR_Savanas	50.17	0.92	7.699999809
-Sample_10	-23.8969	-46.2078	Atlanctic_forest	83.07	3.03	11.69999981
-Sample_11	-23.9017	-45.2508	Atlanctic_forest	3.92	1.67	9.899999619
-Sample_12	-25.0839	-47.9506	Atlanctic_forest	88.23	3.08	9.199999809
-Sample_13	-23.8468	-46.4159	Atlanctic_forest	5.42	5.4	11.89999962
-Sample_14	-23.8658	-46.4375	Atlanctic_forest	72.91	4.6	10.69999981
-Sample_15	-11.6833	-55.8358	Amazon	88.23	3.89	11.10000038
-Sample_17	-11.6833	-55.8358	Amazon	90.21	4	11.10000038
-Sample_18	-11.7413	-56.2533	Amazon	92.5	4.04	9.899999619
-Sample_19	-11.6833	-55.8358	Amazon	92.73	4.04	11.10000038
-Sample_2	-19.51	-43.48	BR_Savanas	64.9	0.85	10
-Sample_20	-11.7413	-56.2533	Amazon	94.09	4.22	9.899999619
-Sample_21	-11.6833	-55.8358	Amazon	94.33	4.6	11.10000038
-Sample_22	-11.7413	-56.2533	Amazon	116.23	4.68	9.899999619
-Sample_23	-8.8338	-42.5537	Caatinga	53.58	1.31	4.300000191
-Sample_24	-9.2236	-41.0865	Caatinga	60.16	2.83	3.200000048
-Sample_25	-6.7123	-38.2523	Caatinga	27.24	3.13	4.699999809
-Sample_26	-9.2236	-41.0865	Caatinga	26.64	2.88	3.200000048
-Sample_27	-8.8338	-42.5537	Caatinga	19.1	0.08	4.300000191
-Sample_28	-6.7123	-38.2523	Caatinga	21.5	2.35	4.699999809
-Sample_29	-15.1958	-54.0592	BR_Savanas	54.74	0.28	8.199999809
-Sample_3	-15.5558	-47.5102	BR_Savanas	62.53	2.07	8.899999619
-Sample_30	-15.1958	-54.0581	BR_Savanas	41.34	0.35	8.199999809
-Sample_31	-14.3606	-54.3569	BR_Savanas	38.98	5.51	7.300000191
-Sample_32	-13.3658	-54.9072	BR_Savanas	53.99	1.15	7.5
-Sample_33	-13.3658	-54.9067	BR_Savanas	9.02	4.68	8.399999619
-Sample_34	-13.3658	-54.9061	BR_Savanas	42.67	7.16	8.399999619
-Sample_35	-15.1958	-54.0586	BR_Savanas	41.74	1.59	8.199999809
-Sample_36	-14.3272	-54.3742	BR_Savanas	23.17	5.85	5.800000191
-Sample_37	-14.3272	-54.3736	BR_Savanas	26.69	1.11	5.800000191
-Sample_38	-14.3272	-54.3747	BR_Savanas	44.12	2.83	5.800000191
-Sample_39	-13.3825	-54.94	BR_Savanas	12.27	3.89	8.899999619
-Sample_4	-22.41	-47.38	Atlanctic_forest	102.12	5.21	7.599999905
-Sample_40	-13.3825	-54.9394	BR_Savanas	17.43	5.5	8.899999619
-Sample_41	-13.3825	-54.9406	BR_Savanas	33.12	0.88	8.899999619
-Sample_42	-14.3606	-54.3575	BR_Savanas	76.31	2.33	7.300000191
-Sample_43	-14.3606	-54.3581	BR_Savanas	57.13	5.08	6.900000095
-Sample_44	-25.0833	-50.15	Atlanctic_forest	98.83	2.83	10
-Sample_45	-10.65	-37	Atlanctic_forest	21.28	2.84	6.800000191
-Sample_46	-23.3	-51.2	Atlanctic_forest	67.92	2.87	8.899999619
-Sample_47	-9.0744	-40.3301	Caatinga	13.68	0.54	3.599999905
-Sample_48	-14.235	-51.9253	BR_Savanas	30.46	3.11	6.199999809
-Sample_5	-19.8635	-43.8056	Atlanctic_forest	104.26	2.83	8.5
-Sample_50	-15.6094	-47.7433	BR_Savanas	47.12	1.29	10.69999981
-Sample_51	-22.41	-47.38	Atlanctic_forest	75.6	2.88	7.599999905
-Sample_52	-15.1012	-49.0107	BR_Savanas	73.1	4.04	8.300000191
-Sample_53	-15.1086	-49.0208	BR_Savanas	22.3	2.56	8.399999619
-Sample_54	-15.1012	-49.0107	BR_Savanas	36.54	4.04	8.300000191
-Sample_55	-22.2752	-54.8169	BR_Savanas	52.35	2.62	8.300000191
-Sample_56	-22.41	-47.38	Atlanctic_forest	101.68	2.98	7.599999905
-Sample_57	-21.32	-48.1284	BR_Savanas	98.61	3.5	7.699999809
-Sample_58	-7.1649	-39.2617	Caatinga	40.56	5.21	4.300000191
-Sample_59	-7.1216	-39.1931	Caatinga	66.88	3.34	4.900000095
-Sample_6	-19.8612	-43.8074	Atlanctic_forest	82.13	2.84	7.699999809
-Sample_60	-7.1445	-39.1859	Caatinga	8.1	3.12	5
-Sample_61	-7.1297	-39.1814	Caatinga	12.71	2.25	4.800000191
-Sample_62	-2.8566	-54.9579	Amazon	115.12	4.82	8.300000191
-Sample_63	-3.2633	-54.8933	Amazon	53.11	4.87	8.800000191
-Sample_64	-3.1314	-54.9578	Amazon	88.23	4.89	8.300000191
-Sample_65	-3.2633	-54.8933	Amazon	90.21	5.08	8.800000191
-Sample_66	-3.1314	-54.9578	Amazon	92.5	5.21	8.300000191
-Sample_67	-3.2633	-54.8933	Amazon	92.73	3.89	8.800000191
-Sample_68	-3.1314	-54.9578	Amazon	94.09	4	8.300000191
-Sample_69	-3.2633	-54.8933	Amazon	94.33	4.04	8.800000191
-Sample_7	-13.3679	-39.0561	Atlanctic_forest	44.61	2.87	9.899999619
-Sample_70	-3.1314	-54.9578	Amazon	110.81	4.04	8.300000191
-Sample_71	-3.2633	-54.8933	Amazon	88.23	4.22	8.800000191
-Sample_72	-3.1314	-54.9578	Amazon	90.21	4.6	8.300000191
-Sample_73	-2.8566	-54.9579	Amazon	92.5	4.68	8.300000191
-Sample_74	-3.1314	-54.9578	Amazon	92.73	4.82	8.300000191
-Sample_75	-2.8566	-54.9579	Amazon	94.09	4.87	8.300000191
-Sample_76	-4.1012	-54.9095	Amazon	94.33	4.89	9
-Sample_77	-19.8339	-43.7354	Atlanctic_forest	70.54	3.03	8.600000381
-Sample_78	-10.65	-37	Atlanctic_forest	48.55	1.43	6.800000191
-Sample_79	-10.17	-62.8	Amazon	100.62	5.08	7.099999905
-Sample_8	-19.8615	-43.7904	Atlanctic_forest	22.07	2.88	7.699999809
-Sample_80	-22.2752	-54.8169	BR_Savanas	46.99	0.8	8.300000191
-Sample_81	-2.856638889	-54.95788889	Amazon	71.97	5.21	8.300000191
-Sample_82	-3.263305556	-54.89333333	Amazon	108.06	3.89	8.800000191
-Sample_83	-3.131361111	-54.95780556	Amazon	94.09	4	8.300000191
-Sample_84	-2.856638889	-54.95788889	Amazon	38.23	4.04	8.300000191
-Sample_85	-3.263305556	-54.89333333	Amazon	20.78	4.04	8.800000191
-Sample_86	-3.131361111	-54.95780556	Amazon	94.33	4.22	8.300000191
-Sample_87	-11.68193611	-55.835775	Amazon	32.69	4.6	12.30000019
-Sample_88	-11.68338056	-55.83431944	Amazon	92.5	4.68	12.19999981
-Sample_89	-11.68474444	-55.83709444	Amazon	32.56	4.82	11.5
-Sample_9	-23.9039	-46.4298	Atlanctic_forest	164.3	2.98	10.80000019
-Sample_90	-11.71720556	-55.79485556	Amazon	92.73	4.87	11.10000038
-Sample_91	-14.32722222	-54.37416667	BR_Savanas	97.8	0.1	5.800000191
-Sample_92	-13.3825	-54.94	BR_Savanas	88.54	2.4	8.899999619
-Sample_93	-30.33333333	-54.25	Pampa	29.66	3.73	7.300000191
-Sample_94	-29.75	-53.75	Pampa	100.68	0.99	8.600000381
