--- new_samps.txt
@@ -1,95 +1,0 @@
-	Samples	Latitude	Longitude	Biome	Average.precip	Alpha_div	Org.Matter
-1	Sample_1	-19.56	-43.75	Cerrado	132.37	0.92	5.98
-2	Sample_2	-19.51	-43.48	Cerrado	64.9	0.85	7.42
-3	Sample_3	-15.5558	-47.5102	Cerrado	62.53	2.07	1.48
-4	Sample_4	-22.41	-47.38	Atlanctic_forest	102.12	3.47	9.19
-5	Sample_5	-19.8635	-43.8056	Atlanctic_forest	104.26	3.79	4.78
-6	Sample_6	-19.8612	-43.8074	Atlanctic_forest	82.13	1.78	5.69
-7	Sample_7	-13.3679	-39.0561	Atlanctic_forest	44.61	0.99	12.79
-8	Sample_8	-19.8615	-43.7904	Atlanctic_forest	22.07	3.14	8.13
-9	Sample_9	-23.9039	-46.4298	Atlanctic_forest	164.3	0.99	16.99
-10	Sample_10	-23.8969	-46.2078	Atlanctic_forest	83.07	3.21	6.06
-11	Sample_11	-23.9017	-45.2508	Atlanctic_forest	3.92	1.67	3.69
-12	Sample_12	-25.0839	-47.9506	Atlanctic_forest	88.23	3.08	5.32
-13	Sample_13	-23.8468	-46.4159	Atlanctic_forest	5.42	0.24	11.09
-14	Sample_14	-23.8658	-46.4375	Atlanctic_forest	72.91	4.6	8.64
-15	Sample_15	-11.6833	-55.8358	Amazon	11.69	3.47	9.19
-16	Sample_16	-26.733	-54.683	Argentina	80.67	1.08	11.42
-17	Sample_17	-11.6833	-55.8358	Amazon	59.36	1.34	13.94
-18	Sample_18	-11.7413	-56.2533	Amazon	68.94	1.88	4.53
-19	Sample_19	-11.6833	-55.8358	Amazon	44.4	1.54	5.02
-20	Sample_20	-11.7413	-56.2533	Amazon	64.98	0.69	6.63
-21	Sample_21	-11.6833	-55.8358	Amazon	33.59	2.58	7.69
-22	Sample_22	-11.7413	-56.2533	Amazon	116.23	4.87	5.58
-23	Sample_23	-8.8338	-42.5537	Caatinga	53.58	1.31	12.09
-24	Sample_24	-9.2236	-41.0865	Caatinga	60.16	2.83	7.97
-25	Sample_25	-6.7123	-38.2523	Caatinga	77.24	3.13	2.7
-26	Sample_26	-9.2236	-41.0865	Caatinga	26.64	2.88	9.42
-27	Sample_27	-8.8338	-42.5537	Caatinga	19.1	0.08	9.38
-28	Sample_28	-6.7123	-38.2523	Caatinga	51.5	2.35	14.78
-29	Sample_29	-15.1958	-54.0592	Cerrado	54.74	0.28	2.43
-30	Sample_30	-15.1958	-54.0581	Cerrado	41.34	0.35	0.97
-31	Sample_31	-14.3606	-54.3569	Cerrado	38.98	5.51	8.82
-32	Sample_32	-13.3658	-54.9072	Cerrado	53.99	1.15	6.79
-33	Sample_33	-13.3658	-54.9067	Cerrado	9.02	4.68	10.34
-34	Sample_34	-13.3658	-54.9061	Cerrado	82.67	7.16	14.2
-35	Sample_35	-15.1958	-54.0586	Cerrado	69.74	1.59	9.03
-36	Sample_36	-14.3272	-54.3742	Cerrado	54.16	5.85	9.47
-37	Sample_37	-14.3272	-54.3736	Cerrado	26.69	1.11	11.11
-38	Sample_38	-14.3272	-54.3747	Cerrado	78.84	2.83	3.63
-39	Sample_39	-13.3825	-54.94	Cerrado	12.27	3.89	8.46
-40	Sample_40	-13.3825	-54.9394	Cerrado	73.43	5.5	7.24
-41	Sample_41	-13.3825	-54.9406	Cerrado	82.41	0.88	11.9
-42	Sample_42	-14.3606	-54.3575	Cerrado	76.31	2.33	7.02
-43	Sample_43	-14.3606	-54.3581	Cerrado	57.13	5.08	4.02
-44	Sample_44	-25.0833	-50.15	Atlanctic_forest	98.83	0.57	9.83
-45	Sample_45	-10.65	-37	Atlanctic_forest	21.28	2.87	16.69
-46	Sample_46	-23.3	-51.2	Atlanctic_forest	67.92	2.98	8.09
-47	Sample_47	-9.0744	-40.3301	Caatinga	64.68	0.54	7.56
-48	Sample_48	-14.235	-51.9253	Cerrado	30.46	3.11	4.33
-49	Sample_49	-26.733	-54.683	Argentina	8.21	4	4.87
-50	Sample_50	-15.6094	-47.7433	Cerrado	47.12	1.29	2.77
-51	Sample_51	-22.41	-47.38	Atlanctic_forest	75.6	1.96	7.09
-52	Sample_52	-15.1012	-49.0107	Cerrado	73.1	4.04	7.83
-53	Sample_53	-15.1086	-49.0208	Cerrado	2.09	2.56	9.02
-54	Sample_54	-15.1012	-49.0107	Cerrado	36.54	4.04	8.21
-55	Sample_55	-22.2752	-54.8169	Cerrado	109.35	2.62	4.17
-56	Sample_56	-22.41	-47.38	Atlanctic_forest	101.68	0.94	6.88
-57	Sample_57	-21.32	-48.1284	Cerrado	98.61	3.5	5.71
-58	Sample_58	-7.1649	-39.2617	Caatinga	40.56	5.21	10.39
-59	Sample_59	-7.1216	-39.1931	Caatinga	96.88	3.34	16.15
-60	Sample_60	-7.1445	-39.1859	Caatinga	54.11	3.12	7.72
-61	Sample_61	-7.1297	-39.1814	Caatinga	33.71	2.25	10.71
-62	Sample_62	-2.8566	-54.9579	Amazon	115.12	2.13	4.53
-63	Sample_63	-3.2633	-54.8933	Amazon	53.11	0.93	10.7
-64	Sample_64	-3.1314	-54.9578	Amazon	19.39	3.03	12.53
-65	Sample_65	-3.2633	-54.8933	Amazon	51.53	3.8	7.56
-66	Sample_66	-3.1314	-54.9578	Amazon	90.21	0.57	15.64
-67	Sample_67	-3.2633	-54.8933	Amazon	31.59	1.21	11.56
-68	Sample_68	-3.1314	-54.9578	Amazon	64.31	1.2	11.63
-69	Sample_69	-3.2633	-54.8933	Amazon	111.04	4.82	8.18
-70	Sample_70	-3.1314	-54.9578	Amazon	110.81	2.45	11.56
-71	Sample_71	-3.2633	-54.8933	Amazon	33.2	1.82	7.7
-72	Sample_72	-3.1314	-54.9578	Amazon	50.34	1.82	11.81
-73	Sample_73	-2.8566	-54.9579	Amazon	79.92	0.3	16.78
-74	Sample_74	-3.1314	-54.9578	Amazon	19.21	2.61	7.89
-75	Sample_75	-2.8566	-54.9579	Amazon	21.08	5.65	7.05
-76	Sample_76	-4.1012	-54.9095	Amazon	27.25	1.35	9.21
-77	Sample_77	-19.8339	-43.7354	Atlanctic_forest	70.54	6.81	9.76
-78	Sample_78	-10.65	-37	Atlanctic_forest	78.55	1.43	5.29
-79	Sample_79	-10.17	-62.8	Amazon	100.62	6.89	10.39
-80	Sample_80	-22.2752	-54.8169	Cerrado	46.99	0.8	7.29
-81	Sample_81	-2.856638889	-54.95788889	Amazon	71.97	4.89	8.56
-82	Sample_82	-3.263305556	-54.89333333	Amazon	108.06	0.09	12.28
-83	Sample_83	-3.131361111	-54.95780556	Amazon	94.09	6.94	1.44
-84	Sample_84	-2.856638889	-54.95788889	Amazon	38.23	5.61	8.47
-85	Sample_85	-3.263305556	-54.89333333	Amazon	20.78	4.22	5.37
-86	Sample_86	-3.131361111	-54.95780556	Amazon	94.33	1.24	6.41
-87	Sample_87	-11.68193611	-55.835775	Amazon	32.69	6.29	3.93
-88	Sample_88	-11.68338056	-55.83431944	Amazon	92.5	3.44	5.03
-89	Sample_89	-11.68474444	-55.83709444	Amazon	32.56	2.84	4.37
-90	Sample_90	-11.71720556	-55.79485556	Amazon	92.73	0.67	7.18
-91	Sample_91	-14.32722222	-54.37416667	Cerrado	97.8	0.1	5.63
-92	Sample_92	-13.3825	-54.94	Cerrado	110.54	2.4	7.47
-93	Sample_93	-30.33333333	-54.25	Pampa	29.66	3.73	7.61
-94	Sample_94	-29.75	-53.75	Pampa	100.68	0.99	8.54
