+++ resources/Samples/New_samples.txt
@@ -1,0 +1,95 @@
+Samples	Latitude	Longitude	Biome
+Sample_1	-19.56	-43.75	Cerrado
+Sample_2	-19.51	-43.48	Cerrado
+Sample_3	-15.5558	-47.5102	Cerrado
+Sample_4	-22.41	-47.38	Atlanctic_forest
+Sample_5	-19.8635	-43.8056	Atlanctic_forest
+Sample_6	-19.8612	-43.8074	Atlanctic_forest
+Sample_7	-13.3679	-39.0561	Atlanctic_forest
+Sample_8	-19.8615	-43.7904	Atlanctic_forest
+Sample_9	-23.9039	-46.4298	Atlanctic_forest
+Sample_10	-23.8969	-46.2078	Atlanctic_forest
+Sample_11	-23.9017	-45.2508	Atlanctic_forest
+Sample_12	-25.0839	-47.9506	Atlanctic_forest
+Sample_13	-23.8468	-46.4159	Atlanctic_forest
+Sample_14	-23.8658	-46.4375	Atlanctic_forest
+Sample_15	-11.6833	-55.8358	Amazon
+Sample_16	-26.733	-54.683	Argentina
+Sample_17	-11.6833	-55.8358	Amazon
+Sample_18	-11.7413	-56.2533	Amazon
+Sample_19	-11.6833	-55.8358	Amazon
+Sample_20	-11.7413	-56.2533	Amazon
+Sample_21	-11.6833	-55.8358	Amazon
+Sample_22	-11.7413	-56.2533	Amazon
+Sample_23	-8.8338	-42.5537	Caatinga
+Sample_24	-9.2236	-41.0865	Caatinga
+Sample_25	-6.7123	-38.2523	Caatinga
+Sample_26	-9.2236	-41.0865	Caatinga
+Sample_27	-8.8338	-42.5537	Caatinga
+Sample_28	-6.7123	-38.2523	Caatinga
+Sample_29	-15.1958	-54.0592	Cerrado
+Sample_30	-15.1958	-54.0581	Cerrado
+Sample_31	-14.3606	-54.3569	Cerrado
+Sample_32	-13.3658	-54.9072	Cerrado
+Sample_33	-13.3658	-54.9067	Cerrado
+Sample_34	-13.3658	-54.9061	Cerrado
+Sample_35	-15.1958	-54.0586	Cerrado
+Sample_36	-14.3272	-54.3742	Cerrado
+Sample_37	-14.3272	-54.3736	Cerrado
+Sample_38	-14.3272	-54.3747	Cerrado
+Sample_39	-13.3825	-54.94	Cerrado
+Sample_40	-13.3825	-54.9394	Cerrado
+Sample_41	-13.3825	-54.9406	Cerrado
+Sample_42	-14.3606	-54.3575	Cerrado
+Sample_43	-14.3606	-54.3581	Cerrado
+Sample_44	-25.0833	-50.15	Atlanctic_forest
+Sample_45	-10.65	-37	Atlanctic_forest
+Sample_46	-23.3	-51.2	Atlanctic_forest
+Sample_47	-9.0744	-40.3301	Caatinga
+Sample_48	-14.235	-51.9253	Cerrado
+Sample_49	-26.733	-54.683	Argentina
+Sample_50	-15.6094	-47.7433	Cerrado
+Sample_51	-22.41	-47.38	Atlanctic_forest
+Sample_52	-15.1012	-49.0107	Cerrado
+Sample_53	-15.1086	-49.0208	Cerrado
+Sample_54	-15.1012	-49.0107	Cerrado
+Sample_55	-22.2752	-54.8169	Cerrado
+Sample_56	-22.41	-47.38	Atlanctic_forest
+Sample_57	-21.32	-48.1284	Cerrado
+Sample_58	-7.1649	-39.2617	Caatinga
+Sample_59	-7.1216	-39.1931	Caatinga
+Sample_60	-7.1445	-39.1859	Caatinga
+Sample_61	-7.1297	-39.1814	Caatinga
+Sample_62	-2.8566	-54.9579	Amazon
+Sample_63	-3.2633	-54.8933	Amazon
+Sample_64	-3.1314	-54.9578	Amazon
+Sample_65	-3.2633	-54.8933	Amazon
+Sample_66	-3.1314	-54.9578	Amazon
+Sample_67	-3.2633	-54.8933	Amazon
+Sample_68	-3.1314	-54.9578	Amazon
+Sample_69	-3.2633	-54.8933	Amazon
+Sample_70	-3.1314	-54.9578	Amazon
+Sample_71	-3.2633	-54.8933	Amazon
+Sample_72	-3.1314	-54.9578	Amazon
+Sample_73	-2.8566	-54.9579	Amazon
+Sample_74	-3.1314	-54.9578	Amazon
+Sample_75	-2.8566	-54.9579	Amazon
+Sample_76	-4.1012	-54.9095	Amazon
+Sample_77	-19.8339	-43.7354	Atlanctic_forest
+Sample_78	-10.65	-37	Atlanctic_forest
+Sample_79	-10.17	-62.8	Amazon
+Sample_80	-22.2752	-54.8169	Cerrado
+Sample_81	-2.856638889	-54.95788889	Amazon
+Sample_82	-3.263305556	-54.89333333	Amazon
+Sample_83	-3.131361111	-54.95780556	Amazon
+Sample_84	-2.856638889	-54.95788889	Amazon
+Sample_85	-3.263305556	-54.89333333	Amazon
+Sample_86	-3.131361111	-54.95780556	Amazon
+Sample_87	-11.68193611	-55.835775	Amazon
+Sample_88	-11.68338056	-55.83431944	Amazon
+Sample_89	-11.68474444	-55.83709444	Amazon
+Sample_90	-11.71720556	-55.79485556	Amazon
+Sample_91	-14.32722222	-54.37416667	Cerrado
+Sample_92	-13.3825	-54.94	Cerrado
+Sample_93	-30.33333333	-54.25	Pampa
+Sample_94	-29.75	-53.75	Pampa
