+++ resources/add_layout_default_info_to_html.sh
