+++ resources/index.Rmd
@@ -1,0 +1,50 @@
+---
+title: 
+
+output: html_document
+---
+
+```{r setup, include=FALSE}
+knitr::opts_chunk$set(echo = TRUE)
+```
+
+```{=html}
+<div class="list_container">
+  <ul class="menu">
+    <li><a href="{{ site.github.tar_url }}">Download <strong>TAR Ball</strong></a></li>
+    <li><a href="{{ site.github.zip_url }}">Download <strong>ZIP File</strong></a></li>
+    <li><a href="{{ site.github.repository_url }}">View On <strong>GitHub</strong></a></li>
+    <li><a href="under_construction.html">Get map by <strong>Taxa</strong></a></li>
+    <li><a href="under_construction.html">Precipitation <strong>Gradients</strong></a></li>
+    <li><a href="under_construction.html">Macronutrient <strong>Effects</strong></a></li>
+  </ul>
+</div>
+```
+### Explore the soil samples containing the microbial diversity of the Brazilian soils.
+
+#### Select the samples of interest and explore the results
+
+
+#### Interactive plots
+##### Use the map below to group and study soil microbiome features.
+```{r, echo=FALSE, message = FALSE, warning = FALSE}
+library(leaflet)
+soil_samples_df <- read.delim(file = "Samples/Sample_metadata.txt")
+
+# feed shared object
+data_ct <- crosstalk::SharedData$new(soil_samples_df[,c(1,5,6,7,11,12,13)])
+
+crosstalk::bscols(
+  widths = c(6,6,8), device = "xs",
+  
+  leaflet(data_ct, width = "100%", height = 300) %>%
+    addTiles() %>%
+    addMarkers(),
+  
+  d3scatter::d3scatter(data = data_ct, ~Org.Matter, ~Alpha_div, ~Biome,
+                         width = "100%", height = 300),
+
+  DT::datatable(data_ct,rownames = FALSE, filter = 'top',
+                width = 600, height = 600)
+  )
+```
