+++ resources/index_regions.Rmd
@@ -1,0 +1,73 @@
+---
+title: 
+
+output: html_document
+---
+
+```{r setup, include=FALSE}
+knitr::opts_chunk$set(echo = TRUE)
+```
+
+```{=html}
+<ul class="menu">
+          <li><a href="{{ site.github.zip_url }}">Download <strong>ZIP File</strong></a></li>
+          <li><a href="{{ site.github.tar_url }}">Download <strong>TAR Ball</strong></a></li>
+          <li><a href="{{ site.github.repository_url }}">View On <strong>GitHub</strong></a></li>
+          <li><a href="{{ site.github.repository_url }}">Get map by <strong>Taxa</strong></a></li>
+          <li><a href="{{ site.github.repository_url }}">Precipitation <strong>Gradients</strong></a></li>
+          <li><a href="{{ site.github.repository_url }}">Macronutrient <strong>Effects</strong></a></li>
+        </ul>
+```
+### Explore the soil samples containing the microbial diversity of the Brazilian soils.
+
+#### Select the samples of interest and explore the results
+
+
+#### Interactive plots
+##### Use the map below to group and study soil microbiome features.
+```{r, echo=FALSE, message = FALSE, warning = FALSE}
+library(leaflet)
+sf_object <- readRDS(file = "../../download_data/biomes_2019_sf_bject.RDS")
+soil_samples_df <- read.delim(file = "Samples/new_samps_metadata.txt")
+
+## REORDER
+soil_samples_df <- soil_samples_df[c(1,12,23,34,45,56,67,78,89,
+                                     c(2:11),c(13:22),c(24:33),
+                                     c(35:44),c(46:55),c(57:66),
+                                     c(68:77),c(79:88),c(90:94)),]
+
+colors = c("darkgreen","yellow","orange","brown","gray","#008BBD", "#4A6FE3")
+
+# subset coastal systems
+#sf_object_no_coastal_sys <- sf_object[!(sf_object$name_biome %in% "Sistema Costeiro"), ]
+
+# Biome Info
+biome_sampling_site_info <- c(paste("Biome: Amazon Forest", "<br>", "Samples sites: 33"),
+                              paste("Biome: Caatinga (Unique Brazilian Xerophyle System", "<br>", "Samples sites: 11"),
+                              paste("Biome: Cerrado (Brazilian Savanas)", "<br>", "Samples sites: 28"),
+                              paste("Biome: Actlantic Forest", "<br>", "Samples sites: 18"),
+                              paste("Biome: Pampa", "<br>", "Samples sites: 2"),
+                              paste("Biome: Pantanal (Wetlands)", "<br>", "Samples sites: 0"),
+                              paste("Biome: Coastal System", "<br>", "Samples sites: 0"))
+
+# feed shared object
+data_ct <- crosstalk::SharedData$new(soil_samples_df)
+
+crosstalk::bscols(
+  widths = c(6,6,8), device = "xs",
+  
+  leaflet(data_ct, width = "100%", height = 300) %>%
+    addPolygons(data = sf_object$geom, 
+                color = colors, 
+                stroke = FALSE, 
+                popup = biome_sampling_site_info) %>%
+    addTiles() %>%
+    addCircles(color = "brown"),
+  
+  d3scatter::d3scatter(data = data_ct, ~Org.Matter, ~Alpha_div, ~Biome,
+                       width = "100%", height = 300),
+  
+  DT::datatable(data_ct,rownames = FALSE, filter = 'none',
+                width = 600, height = 600)
+)
+```
