+++ resources/scripts_data_manipulation/download_mgm_files.bash
