+++ resources/soils_map_files/leaflet-1.3.1/images/marker-shadow.png
